pyproject.toml
[project]
name = "tg-agent-cli"
version = "0.1.0"
description = "A simplified agent CLI framework for running LLM agents with tool calling"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.7,<3",
    "openai>=2.15,<3",
    "httpx>=0.27,<0.28",
    "markdown-it-py>=2.2,<3",
    "prompt-toolkit>=3.0,<4",
    "packaging>=23,<25",
    "rich>=13,<14",
    "python-dotenv>=1,<2",
    "PyYAML>=6,<7",
    "pygments>=2,<3",
    "anyio>=3.0.0",
    "python-telegram-bot>=22.6",
    "fastapi>=0.135.1",
    "uvicorn>=0.41.0",
    "sqlalchemy>=2.0,<3",
    "pymysql[rsa]>=1.1,<2",
    "pytz>=2024.1,<2027",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4,<9",
    "pytest-asyncio>=0.21,<1",
    "black>=24.8,<27",
    "ruff>=0.6,<1",
]
server = [
    "fastapi>=0.116,<1",
    "uvicorn>=0.35,<1",
]

[project.scripts]
tg-agent = "claude_code:cli_main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel.force-include]
"agent_core" = "tg_agent_cli/agent_core"
"cli" = "tg_agent_cli/cli"
"tools" = "tg_agent_cli/tools"
"tg_mem" = "tg_agent_cli/tg_mem"
"claude_code.py" = "tg_agent_cli/claude_code.py"
"config" = "tg_agent_cli/config"
"ralph_init.py" = "tg_agent_cli/ralph_init.py"
"ralph_loop.py" = "tg_agent_cli/ralph_loop.py"
"template" = "tg_agent_cli/template"
".devagent" = "tg_agent_cli/.devagent"
".env.example" = "tg_agent_cli/.env.example"
"README.md" = "tg_agent_cli/README.md"
"plugins" = "tg_agent_cli/plugins"
"tg_agent_cli.pth" = "tg_agent_cli.pth"

[tool.hatch.build.targets.sdist]
include = [
    "agent_core/**",
    "cli/**",
    "tools/**",
    "tg_mem/**",
    "claude_code.py",
    "config/**",
    "ralph_init.py",
    "ralph_loop.py",
    "template/**",
    ".devagent/**",
    ".env.example",
    "README.md",
    "plugins/**",
    "tg_agent_cli.pth",
    "pyproject.toml",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W"]

[tool.poetry]
package-mode = false
